Improve Dockerfile
This commit cuts down the final size of the Docker image from
approx. 18 GB to 9 GB, so about half. (This is still huge, but
better.) It also performs the download and installation of the
models first, so that this step is cached in subsequent builds
and does not need to be redone every time the application
changes. As a consequence, the cache needs to be invalidated
manually in case the model data changes on the remote site.